+++ spec/sandal/sig/rs_spec.rb
@@ -7,8 +7,8 @@ describe Sandal::Sig::RS256 do
     private_key = OpenSSL::PKey::RSA.generate(2048)
     signer = Sandal::Sig::RS256.new(private_key)
     signature = signer.sign(data)
-    verifier = Sandal::Sig::RS256.new(private_key.public_key)
-    verifier.verify(signature, data).should == true
+    validator = Sandal::Sig::RS256.new(private_key.public_key)
+    validator.valid?(signature, data).should == true
   end
 end
 
@@ -18,8 +18,8 @@ describe Sandal::Sig::RS384 do
     private_key = OpenSSL::PKey::RSA.generate(2048)
     signer = Sandal::Sig::RS384.new(private_key)
     signature = signer.sign(data)
-    verifier = Sandal::Sig::RS384.new(private_key.public_key)
-    verifier.verify(signature, data).should == true
+    validator = Sandal::Sig::RS384.new(private_key.public_key)
+    validator.valid?(signature, data).should == true
   end
 end
 
@@ -29,7 +29,7 @@ describe Sandal::Sig::RS512 do
     private_key = OpenSSL::PKey::RSA.generate(2048)
     signer = Sandal::Sig::RS512.new(private_key)
     signature = signer.sign(data)
-    verifier = Sandal::Sig::RS512.new(private_key.public_key)
-    verifier.verify(signature, data).should == true
+    validator = Sandal::Sig::RS512.new(private_key.public_key)
+    validator.valid?(signature, data).should == true
   end
 end