+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 Features:
 
 - Changed from json to multi_json.
+- Added a new Claims module which can be mixed in.
+
+Bug fixes:
+
+- Base64 decoding now ensures there is no padding before decoding.
 
 ## 0.1.0 (30 March 2013)